requirements.txt

playwright
requests
pillow